requirements.txt
Django==2.2
django-maintenance-mode==0.16.0
djangorestframework==3.12.4
greenlet==1.1.0
gunicorn==20.1.0
pytz==2021.1
SQLAlchemy==1.4.20
sqlparse==0.4.1
django-cors-headers==3.2.1
